telegram_url_to_screenshot/pyproject.toml
[tool.poetry]
name = "telegram-screenshot-bot"
version = "1.0"
description = ""
authors = ["Dmitriy Dmytruk <dinamo.mutu - Skype>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
python-telegram-bot = "^12.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
